+++ example/web/e2ee.worker.dart.js
@@ -3798,9 +3798,9 @@ a6.b=a6.c=null
 a6.d=n.x
 if(!n.gaZ()||a7.d.length===0){n.z.b3()
 if(n.e.d.r){s=1
-break}c.i(B.l,"enqueing empty frame",null,null)
+break}c.i(B.l,"enqueuing empty frame",null,null)
 b1.enqueue(b0)
-c.i(B.b,"enqueing silent frame",null,null)
+c.i(B.b,"enqueuing silent frame",null,null)
 s=1
 break}b=n.e.d.e
 if(b!=null){a=a7.d
@@ -3821,13 +3821,13 @@ else a6=!1
 if(a6){a6=B.e.aE(a7.d,a7.d.length-1)
 if(0>=a6.length){q=A.e(a6,0)
 s=1
-break}c.i(B.b,"ecodeFunction: skip uncrypted frame, type "+a6[0],null,null)
+break}c.i(B.b,"decodeFunction: skip unencrypted frame, type "+a6[0],null,null)
 e=new A.b2($.cQ())
 e.u(0,new Uint8Array(A.ar(B.e.B(a7.d,0,a7.d.length-a1))))
 n.ap(b0,b1,e)
-c.i(B.l,"ecodeFunction: enqueing silent frame",null,null)
-b1.enqueue(b0)}else c.i(B.b,"ecodeFunction: SIF limit reached, dropping frame",null,null)
-c.i(B.b,"ecodeFunction: enqueing silent frame",null,null)
+c.i(B.l,"decodeFunction: enqueuing silent frame",null,null)
+b1.enqueue(b0)}else c.i(B.b,"decodeFunction: SIF limit reached, dropping frame",null,null)
+c.i(B.b,"decodeFunction: enqueuing silent frame",null,null)
 b1.enqueue(b0)
 s=1
 break}else a.b3()}}p=4
@@ -4566,7 +4566,7 @@ s(A.bT,A.L)})()
 var v={G:typeof self!="undefined"?self:globalThis,typeUniverse:{eC:new Map(),tR:{},eT:{},tPV:{},sEA:[]},mangledGlobalNames:{a:"int",n:"double",aO:"num",a5:"String",V:"bool",y:"Null",r:"List",k:"Object",bs:"Map",q:"JSObject"},mangledNames:{},types:["~()","V(ak)","a_<~>()","~(@)","~(~())","y(@)","y()","~(k,a1)","k?(k?)","~(q,q)","V(aw)","y(q)","V(ah)","@(@)","@(@,a5)","@(a5)","y(~())","y(@,a1)","~(a,@)","y(k,a1)","~(k?,k?)","aX()","~(aB)","a_<y>(q)"],interceptorsByTag:null,leafTags:null,arrayRti:Symbol("$ti")}
 A.id(v.typeUniverse,JSON.parse('{"a6":"al","ct":"al","bE":"al","jx":"an","cf":{"V":[],"o":[]},"bm":{"y":[],"o":[]},"bn":{"q":[]},"al":{"q":[]},"z":{"r":["1"],"l":["1"],"q":[],"f":["1"]},"ce":{"bC":[]},"d6":{"z":["1"],"r":["1"],"l":["1"],"q":[],"f":["1"]},"bf":{"a4":["1"]},"ch":{"n":[],"aO":[]},"bl":{"n":[],"a":[],"aO":[],"o":[]},"cg":{"n":[],"aO":[],"o":[]},"aT":{"a5":[],"fa":[],"o":[]},"b2":{"hm":[]},"bo":{"w":[]},"l":{"f":["1"]},"a7":{"l":["1"],"f":["1"]},"aA":{"a4":["1"]},"a8":{"f":["2"],"f.E":"2"},"bi":{"a8":["1","2"],"l":["2"],"f":["2"],"f.E":"2"},"bt":{"a4":["2"]},"a9":{"a7":["2"],"l":["2"],"f":["2"],"f.E":"2","a7.E":"2"},"aE":{"f":["1"],"f.E":"1"},"bH":{"a4":["1"]},"bA":{"ab":[],"w":[]},"ci":{"w":[]},"cB":{"w":[]},"bU":{"a1":[]},"aj":{"ax":[]},"c6":{"ax":[]},"c7":{"ax":[]},"cy":{"ax":[]},"cw":{"ax":[]},"aR":{"ax":[]},"cv":{"w":[]},"az":{"aC":["1","2"],"f3":["1","2"],"bs":["1","2"]},"bq":{"l":["1"],"f":["1"],"f.E":"1"},"bp":{"a4":["1"]},"aY":{"an":[],"q":[],"bg":[],"o":[]},"an":{"q":[],"bg":[],"o":[]},"bx":{"q":[]},"cL":{"bg":[]},"bu":{"ez":[],"q":[],"o":[]},"C":{"P":["1"],"q":[]},"bv":{"u":["n"],"C":["n"],"r":["n"],"P":["n"],"l":["n"],"q":[],"f":["n"],"L":["n"]},"bw":{"u":["a"],"C":["a"],"r":["a"],"P":["a"],"l":["a"],"q":[],"f":["a"],"L":["a"]},"ck":{"cX":[],"u":["n"],"C":["n"],"r":["n"],"P":["n"],"l":["n"],"q":[],"f":["n"],"L":["n"],"o":[],"u.E":"n"},"cl":{"cY":[],"u":["n"],"C":["n"],"r":["n"],"P":["n"],"l":["n"],"q":[],"f":["n"],"L":["n"],"o":[],"u.E":"n"},"cm":{"d2":[],"u":["a"],"C":["a"],"r":["a"],"P":["a"],"l":["a"],"q":[],"f":["a"],"L":["a"],"o":[],"u.E":"a"},"cn":{"d3":[],"u":["a"],"C":["a"],"r":["a"],"P":["a"],"l":["a"],"q":[],"f":["a"],"L":["a"],"o":[],"u.E":"a"},"co":{"d4":[],"u":["a"],"C":["a"],"r":["a"],"P":["a"],"l":["a"],"q":[],"f":["a"],"L":["a"],"o":[],"u.E":"a"},"cp":{"dl":[],"u":["a"],"C":["a"],"r":["a"],"P":["a"],"l":["a"],"q":[],"f":["a"],"L":["a"],"o":[],"u.E":"a"},"cq":{"dm":[],"u":["a"],"C":["a"],"r":["a"],"P":["a"],"l":["a"],"q":[],"f":["a"],"L":["a"],"o":[],"u.E":"a"},"by":{"dn":[],"u":["a"],"C":["a"],"r":["a"],"P":["a"],"l":["a"],"q":[],"f":["a"],"L":["a"],"o":[],"u.E":"a"},"bz":{"cz":[],"u":["a"],"C":["a"],"r":["a"],"P":["a"],"l":["a"],"q":[],"f":["a"],"L":["a"],"o":[],"u.E":"a"},"cF":{"w":[]},"bW":{"ab":[],"w":[]},"ad":{"b0":["1"],"ap":["1"]},"O":{"w":[]},"b1":{"bJ":["1"],"b5":["1"],"b_":["1"]},"ao":{"bK":["1"],"ad":["1"],"b0":["1"],"ap":["1"]},"aF":{"fh":["1"],"fv":["1"],"ap":["1"]},"bV":{"aF":["1"],"fh":["1"],"fv":["1"],"ap":["1"]},"bI":{"cE":["1"]},"x":{"a_":["1"]},"bJ":{"b5":["1"],"b_":["1"]},"bK":{"ad":["1"],"b0":["1"],"ap":["1"]},"b5":{"b_":["1"]},"bL":{"bM":["1"]},"b3":{"b0":["1"]},"c_":{"fm":[]},"cI":{"c_":[],"fm":[]},"bN":{"aC":["1","2"],"bs":["1","2"]},"b4":{"bN":["1","2"],"aC":["1","2"],"bs":["1","2"]},"bO":{"l":["1"],"f":["1"],"f.E":"1"},"bP":{"a4":["1"]},"aC":{"bs":["1","2"]},"c5":{"av":["r<a>","a5"],"av.S":"r<a>"},"n":{"aO":[]},"a":{"aO":[]},"r":{"l":["1"],"f":["1"]},"a5":{"fa":[]},"c3":{"w":[]},"ab":{"w":[]},"Z":{"w":[]},"aZ":{"w":[]},"cc":{"w":[]},"bF":{"w":[]},"cA":{"w":[]},"aD":{"w":[]},"c8":{"w":[]},"cr":{"w":[]},"bD":{"w":[]},"cK":{"a1":[]},"d4":{"r":["a"],"l":["a"],"f":["a"]},"cz":{"r":["a"],"l":["a"],"f":["a"]},"dn":{"r":["a"],"l":["a"],"f":["a"]},"d2":{"r":["a"],"l":["a"],"f":["a"]},"dl":{"r":["a"],"l":["a"],"f":["a"]},"d3":{"r":["a"],"l":["a"],"f":["a"]},"dm":{"r":["a"],"l":["a"],"f":["a"]},"cX":{"r":["n"],"l":["n"],"f":["n"]},"cY":{"r":["n"],"l":["n"],"f":["n"]}}'))
 A.ic(v.typeUniverse,JSON.parse('{"l":1,"C":1,"bM":1,"c9":2}'))
-var u={o:"Cannot fire new event. Controller is already firing an event",c:"Error handler must accept one Object or one Object and a StackTrace as arguments, and return a value of the returned future's type",r:"[decodeFunction] decryption failed even after ratchting",u:"[ratchedKeyInternal] cannot ratchet anymore",h:"]: lastError != CryptorError.kOk, reset state to kNew",f:"decodeFunction: decryption success, buffer length ",D:"decodeFunction::decryptFrameInternal: decrypted: ",E:"decodeFunction::decryptFrameInternal: ratchetKey: decryption ok, newState: kKeyRatcheted"}
+var u={o:"Cannot fire new event. Controller is already firing an event",c:"Error handler must accept one Object or one Object and a StackTrace as arguments, and return a value of the returned future's type",r:"[decodeFunction] decryption failed even after ratcheting",u:"[ratchetKeyInternal] cannot ratchet anymore",h:"]: lastError != CryptorError.kOk, reset state to kNew",f:"decodeFunction: decryption success, buffer length ",D:"decodeFunction::decryptFrameInternal: decrypted: ",E:"decodeFunction::decryptFrameInternal: ratchetKey: decryption ok, newState: kKeyRatcheted"}
 var t=(function rtii(){var s=A.ba
 return{h:s("@<~>"),b:s("ah"),n:s("O"),B:s("c5"),J:s("bg"),V:s("ez"),D:s("aw"),d:s("l<@>"),C:s("w"),G:s("cX"),q:s("cY"),j:s("ak"),Z:s("ax"),O:s("d2"),k:s("d3"),U:s("d4"),R:s("f<@>"),e:s("f<a>"),s:s("z<a5>"),r:s("z<@>"),t:s("z<a>"),c:s("z<k?>"),u:s("bm"),m:s("q"),g:s("a6"),w:s("P<@>"),x:s("aW"),cK:s("r<@>"),L:s("r<a>"),bG:s("r<aW?>"),cH:s("aB"),I:s("aX"),f:s("bs<@,@>"),a:s("aY"),P:s("y"),K:s("k"),bW:s("cs"),cY:s("jz"),l:s("a1"),N:s("a5"),a4:s("o"),b7:s("ab"),c0:s("dl"),bk:s("dm"),ca:s("dn"),p:s("cz"),cr:s("bE"),_:s("x<@>"),aQ:s("x<a>"),A:s("b4<k?,k?>"),W:s("bV<aB>"),y:s("V"),c1:s("V(k)"),i:s("n"),z:s("@"),bd:s("@()"),v:s("@(k)"),Q:s("@(k,a1)"),S:s("a"),a5:s("bj?"),bc:s("a_<y>?"),b1:s("q?"),aF:s("aW?"),X:s("k?"),T:s("a5?"),E:s("cz?"),F:s("aG<@,@>?"),cG:s("V?"),dd:s("n?"),a3:s("a?"),ae:s("aO?"),Y:s("~()?"),o:s("aO"),H:s("~"),M:s("~()"),bo:s("~(k)"),aD:s("~(k,a1)")}})();(function constants(){var s=hunkHelpers.makeConstList
 B.M=J.cd.prototype