+++ CHANGELOG.md
@@ -62,7 +62,7 @@
 * fix: memory leak visualizer stop android (#831)
 * fix: ensure engine always has correct device id if switching camera fails (#820)
 * fix: Always emit RoomDisconnectedEvent when the reason is clientInitiated. (#821)
-* fix: sif frame enqueing with e2ee (#822)
+* fix: sif frame enqueuing with e2ee (#822)
 * fix: update camera `deviceId` when track is null (#814)
 * feat: Audio Visualizer for Windows/Linux. (#739)